requirements.txt
anyio==4.3.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==2.0.12
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dotenv==1.0.0
python-telegram-bot==21.2
pytz==2024.1
requests==2.26.0
six==1.16.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.11.0
tzlocal==5.2
urllib3==1.26.18
